+++ AlpacaService/Adapter/Contracts/IAlpacaDataService.cs
@@ -1,13 +1,10 @@
-﻿namespace BN.PROJECT.AlpacaService
+﻿namespace BN.PROJECT.AlpacaService;
+
+public interface IAlpacaDataService
 {
-    public interface IAlpacaDataService
-    {
-        Task<List<AlpacaBar>> GetHistoricalBarsBySymbol(string symbol, DateTime startDate, DateTime endDate, BarTimeFrame timeFrame);
-
-        Task<AlpacaBar> GetLatestBarBySymbol(string symbol);
-
-        Task<List<IQuote>> GetQuotesBySymbol(string symbol, DateTime startDate, DateTime endDate);
-
-        Task<List<ITrade>> GetTradesBySymbol(string symbol, DateTime startDate, DateTime endDate);
-    }
+    Task<List<AlpacaBar>> GetHistoricalBarsBySymbol(string symbol, DateTime startDate, DateTime endDate, BarTimeFrame timeFrame);
+    Task<AlpacaBar> GetLatestBarBySymbol(string symbol);
+    Task<List<AlpacaQuote>> GetQuotesBySymbol(string symbol, DateTime startDate, DateTime endDate);
+    Task<AlpacaQuote> GetLatestQuoteBySymbol(string symbol);
+    Task<List<ITrade>> GetTradesBySymbol(string symbol, DateTime startDate, DateTime endDate);
 }