+++ AlpacaService/Domain/Models/SideEnum.cs
@@ -1,0 +1,7 @@
+﻿namespace BN.PROJECT.AlpacaService;
+public enum Side
+{
+    Buy,
+    Sell
+}
+