+++ AlpacaService/Migrations/20240924074809_UserSettingsEmailPK.cs
@@ -1,0 +1,59 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace BN.PROJECT.AlpacaService.Migrations
+{
+    /// <inheritdoc />
+    public partial class UserSettingsEmailPK : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropPrimaryKey(
+                name: "PK_UserSettings",
+                table: "UserSettings");
+
+            migrationBuilder.DropColumn(
+                name: "UserId",
+                table: "UserSettings");
+
+            migrationBuilder.AddColumn<string>(
+                name: "Email",
+                table: "UserSettings",
+                type: "text",
+                nullable: false,
+                defaultValue: "");
+
+            migrationBuilder.AddPrimaryKey(
+                name: "PK_UserSettings",
+                table: "UserSettings",
+                column: "Email");
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropPrimaryKey(
+                name: "PK_UserSettings",
+                table: "UserSettings");
+
+            migrationBuilder.DropColumn(
+                name: "Email",
+                table: "UserSettings");
+
+            migrationBuilder.AddColumn<Guid>(
+                name: "UserId",
+                table: "UserSettings",
+                type: "uuid",
+                nullable: false,
+                defaultValue: new Guid("00000000-0000-0000-0000-000000000000"));
+
+            migrationBuilder.AddPrimaryKey(
+                name: "PK_UserSettings",
+                table: "UserSettings",
+                column: "UserId");
+        }
+    }
+}