+++ AlpacaService/Migrations/20240926090958_Quotes.cs
@@ -1,0 +1,44 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+using Npgsql.EntityFrameworkCore.PostgreSQL.Metadata;
+
+#nullable disable
+
+namespace BN.PROJECT.AlpacaService.Migrations
+{
+    /// <inheritdoc />
+    public partial class Quotes : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.CreateTable(
+                name: "Quotes",
+                columns: table => new
+                {
+                    Id = table.Column<long>(type: "bigint", nullable: false)
+                        .Annotation("Npgsql:ValueGenerationStrategy", NpgsqlValueGenerationStrategy.IdentityByDefaultColumn),
+                    Symbol = table.Column<string>(type: "text", nullable: false),
+                    TimestampUtc = table.Column<DateTime>(type: "timestamp with time zone", nullable: false),
+                    BidExchange = table.Column<string>(type: "text", nullable: false),
+                    AskExchange = table.Column<string>(type: "text", nullable: false),
+                    BidPrice = table.Column<decimal>(type: "numeric", nullable: false),
+                    AskPrice = table.Column<decimal>(type: "numeric", nullable: false),
+                    BidSize = table.Column<decimal>(type: "numeric", nullable: false),
+                    AskSize = table.Column<decimal>(type: "numeric", nullable: false),
+                    Tape = table.Column<string>(type: "text", nullable: false)
+                },
+                constraints: table =>
+                {
+                    table.PrimaryKey("PK_Quotes", x => x.Id);
+                });
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropTable(
+                name: "Quotes");
+        }
+    }
+}